tests/wpt/tests/css/CSS2/borders/discrete-no-interpolation.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/w3c/csswg-drafts/issues/4441">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>
<script>
test_no_interpolation({
  property: 'border-left-style',
  from: 'initial',
  to: 'dotted'
});

test_no_interpolation({
  property: 'border-right-style',
  from: 'initial',
  to: 'dotted'
});

test_no_interpolation({
  property: 'border-top-style',
  from: 'initial',
  to: 'dotted'
});

test_no_interpolation({
  property: 'border-bottom-style',
  from: 'initial',
  to: 'dotted'
});
</script>
